Consolidate Parser and UI packages
