+++ src/main/java/command/DeleteCommand.java
@@ -2,7 +2,7 @@ package command;
 
 import task.Task;
 import task.TaskList;
-import ui.Ui;
+import textual.Ui;
 
 /**
  * Class representing a command to delete an item from the task list.
