+++ src/main/java/command/DoneCommand.java
@@ -1,7 +1,7 @@
 package command;
 
 import task.TaskList;
-import ui.Ui;
+import textual.Ui;
 
 /**
  * Class representing a command to mark an item in the task list as done.
