+++ src/main/java/command/ExitCommand.java
@@ -1,7 +1,7 @@
 package command;
 
 import task.TaskList;
-import ui.Ui;
+import textual.Ui;
 
 /**
  * Class representing a command to exit Duke.
