+++ src/main/java/textual/Ui.java
@@ -1,4 +1,4 @@
-package ui;
+package textual;
 
 import task.TaskList;
